+++ SimTKcommon/Simulation/include/SimTKcommon/internal/State.h
@@ -9,7 +9,7 @@
  * Biological Structures at Stanford, funded under the NIH Roadmap for        *
  * Medical Research, grant U54 GM072970. See https://simtk.org.               *
  *                                                                            *
- * Portions copyright (c) 2005-14 Stanford University and the Authors.        *
+ * Portions copyright (c) 2005-15 Stanford University and the Authors.        *
  * Authors: Michael Sherman                                                   *
  * Contributors: Peter Eastman                                                *
  *                                                                            *
@@ -38,6 +38,7 @@ done in a separate internal class. **/
 #include <cassert>
 #include <algorithm>
 #include <mutex>
+#include <array>
 
 namespace SimTK {
 
@@ -153,13 +154,6 @@ SimTK_DEFINE_UNIQUE_INDEX_TYPE(SystemMultiplierIndex);
 /// Unique integer type for Subsystem-local multiplier indexing
 /// @see SystemMultiplierIndex
 SimTK_DEFINE_UNIQUE_INDEX_TYPE(MultiplierIndex);
-
-/// This is the type to use for Stage version numbers. Whenever any state 
-/// variable is modified, we increment the stage version for the stage(s) that
-/// depend on it. -1 means "unintialized". 0 is never used as a stage version, 
-/// but is allowed as a cache value which is guaranteed never to look valid. 
-typedef int StageVersion;
-
 
 /** This object is intended to contain all state information for a 
 SimTK::System, except topological information which is stored in the %System 
@@ -287,6 +281,9 @@ State();
 /// state variables either, except those associated with the Topology stage.
 State(const State&);
 
+/// The move constructor is very fast. The source object is left empty.
+State(State&& source);
+
 /// Copy assignment has deep copy semantics; that is, \c this %State will 
 /// contain a \e copy of the source, \e not a reference into it. This makes the 
 /// current %State contain a copy of the state information in the source %State,
@@ -294,6 +291,10 @@ State(const State&);
 /// been realized to at least Stage::Model, then we don't copy its state
 /// variables either, except those associated with the Topology stage.
 State& operator=(const State&);
+
+/// Move assignment is very fast. The source object is left in a valid but
+/// undefined condition.
+State& operator=(State&& source);
 
 /// Destruct this %State object and free up the heap space it is using.
 ~State();
@@ -660,15 +661,16 @@ State -- don't delete the object after this call!
 @see allocateLazyCacheEntry(), isCacheValueRealized(), markCacheValueRealized() **/
 inline CacheEntryIndex 
 allocateCacheEntry(SubsystemIndex, Stage earliest, Stage latest,
-                   AbstractValue*) const;
+                   AbstractValue* value, unsigned numExtras=0) const;
 
 /** This is an abbreviation for allocation of a cache entry whose earliest
 and latest Stages are the same. That is, this cache entry is guaranteed to
 be valid if its Subsystem has advanced to the supplied Stage or later, and
 is guaranteed to be invalid below that Stage. **/
 inline CacheEntryIndex 
-allocateCacheEntry(SubsystemIndex sx, Stage g, AbstractValue* v) const
-{   return allocateCacheEntry(sx, g, g, v); }
+allocateCacheEntry(SubsystemIndex sx, Stage stage, 
+                   AbstractValue* value, unsigned numExtras=0) const
+{   return allocateCacheEntry(sx, stage, stage, value, numExtras); }
 
 /** This is an abbreviation for allocation of a lazy cache entry. The 
 \a earliest stage at which this \e can be evaluated is provided; but there is no
@@ -679,9 +681,11 @@ is automatically invalidated when the indicated stage \a earliest is
 invalidated in the State.
 @see allocateCacheEntry(), isCacheValueRealized(), markCacheValueRealized() **/
 inline CacheEntryIndex 
-allocateLazyCacheEntry
-   (SubsystemIndex sx, Stage earliest, AbstractValue* v) const
-{   return allocateCacheEntry(sx, earliest, Stage::Infinity, v); }
+allocateLazyCacheEntry(SubsystemIndex sx, Stage earliest, 
+                       AbstractValue* value, unsigned numExtras=0) const {
+    return allocateCacheEntry(sx, earliest, Stage::Infinity, 
+                              value, numExtras);
+}
 
 /** At what stage was this State when this cache entry was allocated?
 The answer must be Stage::Empty, Stage::Topology, or Stage::Model. **/
@@ -716,6 +720,18 @@ without getting an exception thrown.
 @see allocateCacheEntry(), markCacheValueRealized(), getCacheEntry() **/
 inline bool isCacheValueRealized(SubsystemIndex, CacheEntryIndex) const;
 
+/** (Advanced) Check whether cache entry value is up to date, dealing with
+extra dependencies. The number of extras must match the number provided
+when the cache entry was allocated. If the cache value is up to date, its
+stage version is returned for the convenience of higher-level computations
+that might have a dependency on this one. Otherwise the stage version is
+returned with an invalid value.
+@see markCacheValueRealizedWithExtras() **/
+inline bool isCacheValueRealizedWithExtras
+   (SubsystemIndex, CacheEntryIndex,
+    unsigned numExtras, const StageVersion* extraVersions,
+    StageVersion& cacheValueVersion) const;
+
 /** Mark the value of a particular cache entry as up to date after it has
 been recalculated. This %State's current stage must be at least the
 \a earliest stage as supplied when this cache entry was allocated, and
@@ -730,10 +746,18 @@ a state variable on which it depends.
 @see allocateCacheEntry(), isCacheValueRealized(), getCacheEntry() **/
 inline void markCacheValueRealized(SubsystemIndex, CacheEntryIndex) const;
 
-/** Normally cache entries are invalidated automatically, however this
-method allows manual invalidation of the value of a particular cache 
-entry. After a cache entry has been marked invalid here, 
-isCacheValueRealized() will return false.
+/** (Advanced) Mark a cache entry up to date after recalculating it and
+include version information for extra dependencies. The number of extras must 
+match the number provided when the cache entry was allocated. 
+@see isCacheValueRealizedWithExtras() **/
+inline void markCacheValueRealizedWithExtras
+   (SubsystemIndex, CacheEntryIndex,
+    unsigned numExtras, const StageVersion* extraVersions) const;
+
+/** (Advanced) Normally cache entries are invalidated automatically, however 
+this method allows manual invalidation of the value of a particular cache entry.
+After a cache entry has been marked invalid here, isCacheValueRealized() or
+isCacheValueRealizedWithExtras() will return false.
 @see isCacheValueRealized(), markCacheValueRealized() **/
 inline void markCacheValueNotRealized(SubsystemIndex, CacheEntryIndex) const;
 /**@}**/
@@ -1088,6 +1112,26 @@ was created. This has no effect on the realization level.
 @see getSystemTopologyStageVersion(), System::getSystemTopologyCacheVersion()
 **/
 inline void setSystemTopologyStageVersion(StageVersion topoVersion);
+
+/** (Advanced) Return a StageVersion for time, meaning an integer that is
+incremented whenever time is changed. Will be 1 or greater if time has been
+initialized; otherwise, time is meaningless and may be NaN. **/
+inline StageVersion getTimeStageVersion() const;
+/** (Advanced) Return a StageVersion for q, meaning an integer that is
+incremented whenever any q is changed (or more precisely whenever q or y is
+returned with writable access). Will be 1 or greater if q has been
+allocated. **/
+inline StageVersion getQStageVersion() const;
+/** (Advanced) Return a StageVersion for u, meaning an integer that is
+incremented whenever any u is changed (or more precisely whenever u or y is
+returned with writable access). Will be 1 or greater if u has been
+allocated. **/
+inline StageVersion getUStageVersion() const;
+/** (Advanced) Return a StageVersion for z, meaning an integer that is
+incremented whenever any z is changed (or more precisely whenever z or y is
+returned with writable access). Will be 1 or greater if z has been
+allocated. **/
+inline StageVersion getZStageVersion() const;
 
 /** (Advanced) This is called at the beginning of every integration step to set
 the values of auto-update discrete variables from the values stored in their