+++ CHANGELOG.md
@@ -37,6 +37,9 @@ of the current working directory (thanks to Carmichael Ong). See [Issue #264](ht
 * Added the ability to parallelize forces, robustness/performance improvements
 to ParallelExecutor, mutex state lock.
 [PR #414](https://github.com/simbody/simbody/pull/414).
+* Added move constructor and move assignment to State (very fast).
+* Added "stage version" counters for time, q, u, and z that are incremented
+  whenever one of these changes.
 * (There are more that haven't been added yet)